fix: se cambia logica de midelware y archivos de auth js para bajar peso de middleware y que se pueda desplegar en vercel,
